Refactor SideNavbar component structure by removing the deprecated SideNavbar0 file and updating imports. Adjust localStorage.json formatting for consistency. Clean up axios instance handling by removing unnecessary comments and ensuring proper imports in services. Enhance i18n.js for better language detection and fallback. Update NavigationItems to improve item order and visibility.
