+++ localStorage.json
@@ -1,14 +1,14 @@
 {
     "user": {
-    "userId": 29,
-    "email": "ismailcollege@gmail.com",
-    "name": "Ismail Sherif",
-    "role": "CollegeAdmin",
-    "pictureURL": "https: //emailingsystemapi.runasp.net/Attachments/ProfileImages/20250428-ff86d40b-153e-4cb7-a925-55cc511a5830-20250423-38a96c1c-acda-489e-b0ba-3824b28a8624-2-profile-picture.jpg",
-    "signatureURL": "Empty",
-    "departmentName": null,
-    "collegeName": "Faculty of Computers & AI",
-    "nationalId": "03239847237458"
+        "userId": 29,
+        "email": "ismailcollege@gmail.com",
+        "name": "Ismail Sherif",
+        "role": "CollegeAdmin",
+        "pictureURL": "https: //emailingsystemapi.runasp.net/Attachments/ProfileImages/20250428-ff86d40b-153e-4cb7-a925-55cc511a5830-20250423-38a96c1c-acda-489e-b0ba-3824b28a8624-2-profile-picture.jpg",
+        "signatureURL": "Empty",
+        "departmentName": null,
+        "collegeName": "Faculty of Computers & AI",
+        "nationalId": "03239847237458"
     },
     "isAuthenticated": "true",
     "_persist": {
