+++ src/i18n.js
@@ -7,7 +7,6 @@ import enTranslations from "./locales/en.json";
 // Get browser language
 const getBrowserLanguage = () => {
   const browserLang = navigator.language.split("-")[0]; // Get primary language code
-  // console.log(browserLang);
   return ["en", "ar"].includes(browserLang) ? browserLang : "en"; // Default to 'en' if not supported
 };
 
@@ -16,7 +15,7 @@ i18n.use(initReactI18next).init({
     en: { translation: enTranslations },
     ar: { translation: arTranslations },
   },
-  lng: getBrowserLanguage(), // Use browser language
+  lng: getBrowserLanguage(),
   fallbackLng: "en",
   interpolation: {
     escapeValue: false,
