+++ src/store/index.js
@@ -1,4 +1,4 @@
-// src/store/index.js
+
 import { configureStore } from "@reduxjs/toolkit";
 import {
   persistStore,
